requirements.txt
python-telegram-bot==22.0
python-dotenv==1.1.0
telethon==1.40.0